+++ src/rrcore/queryexecutors/purchase/viewpurchasecart.h
@@ -10,7 +10,7 @@ class ViewPurchaseCart : public PurchaseExecutor
 public:
     static inline const QString COMMAND = QStringLiteral("view_purchase_cart");
 
-    explicit ViewPurchaseCart(qint64 transactionId,
+    explicit ViewPurchaseCart(int transactionId,
                               QObject *receiver);
     QueryResult execute() override;
 };
