+++ src/rrcore/queryexecutors/purchase/viewpurchasedproducts.h
@@ -10,7 +10,7 @@ class ViewPurchasedProducts : public PurchaseExecutor
 public:
     static inline const QString COMMAND = QStringLiteral("view_purchased_products");
 
-    explicit ViewPurchasedProducts(qint64 transactionId,
+    explicit ViewPurchasedProducts(int transactionId,
                                    QObject *receiver);
     QueryResult execute() override;
 };
